+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NEXT RELEASE
-* Store an `updatedAt` value with each data object in the browser storage
+* Store an `updatedAt` value with each data object in the browser storage and use that to determine what data is actually stale
 * Init `TheAlgorithm.loadingStatus` with `"(ready to load)"` during construction so it isn't empty until load is complete
+* Refactor a common fetchTrendingFromAllServers() method wrapper
+* Sort trending toots correctly
 
 ### v0.20.9
 * (Demo App) Make the refresh handling actually work by moving required vars and code into `useEffect()` blocks